glm_ocr/requirements.txt
# GLM-OCR 依赖清单
# Python 3.8+

# PyTorch (根据系统选择安装命令)
# CPU: pip install torch torchvision torchaudio -i https://pypi.tuna.tsinghua.edu.cn/simple
# CUDA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers (需要最新开发版)
# 注意: GLM-OCR 需要从 GitHub 安装最新版
# pip install git+https://github.com/huggingface/transformers.git
# 这里先安装稳定版，setup_env.sh 会升级到开发版
transformers>=5.0.0

# Accelerate (模型加载必需)
accelerate>=0.20.0

# 图像处理
pillow>=9.0.0

# 其他依赖
numpy>=1.21.0
